Totally rewrite all library and Tensor class: - Added dynamic and manual memory control - Added dynamic typing and removed templates - Implemented basic Autograd functions - Added additional tests - Added Arange function for Tensor creation

+++ statkitcpp/core/_tensor/shape.cpp
@@ -1,19 +1,18 @@
 #include "shape.h"
 #include "../errors.h"
-#include <cstdint>
 #include <vector>
 
-bool CheckIndex(int index, uint32_t size) {
+bool CheckIndex(int index, size_t size) {
     return 0 <= index && index < static_cast<int>(size);
 }
 
-bool IsBroadcastable(const std::vector<uint32_t>& shape1,
-                     const std::vector<uint32_t>& shape2) {
+bool IsBroadcastable(const std::vector<size_t>& shape1,
+                     const std::vector<size_t>& shape2) {
     for (int i = 0; i < std::max(shape1.size(), shape2.size()); i++) {
         int index1 = static_cast<int>(shape1.size()) - i - 1;
         int index2 = static_cast<int>(shape2.size()) - i - 1;
-        uint32_t dim1 = CheckIndex(index1, shape1.size()) ? shape1[index1] : 1;
-        uint32_t dim2 = CheckIndex(index2, shape2.size()) ? shape2[index2] : 1;
+        size_t dim1 = CheckIndex(index1, shape1.size()) ? shape1[index1] : 1;
+        size_t dim2 = CheckIndex(index2, shape2.size()) ? shape2[index2] : 1;
         if (dim1 != dim2 && dim1 != 1 && dim2 != 1) {
             return false;
         }
@@ -21,9 +20,9 @@ bool IsBroadcastable(const std::vector<uint32_t>& shape1,
     return true;
 }
 
-std::string ShapeToString(const std::vector<uint32_t>& shape) {
+std::string ShapeToString(const std::vector<size_t>& shape) {
     std::string shape_repr = "(";
-    for (uint32_t i = 0; i < shape.size(); i++) {
+    for (size_t i = 0; i < shape.size(); i++) {
         const auto dim = shape[i];
         shape_repr += std::to_string(dim);
         if (i < shape.size() - 1) {
@@ -34,25 +33,25 @@ std::string ShapeToString(const std::vector<uint32_t>& shape) {
     return shape_repr;
 }
 
-std::vector<uint32_t> BroadcastShapes(const std::vector<uint32_t>& shape1,
-                                      const std::vector<uint32_t>& shape2) {
+std::vector<size_t> BroadcastShapes(const std::vector<size_t>& shape1,
+                                      const std::vector<size_t>& shape2) {
     if (!IsBroadcastable(shape1, shape2)) {
         throw BroadcastError{ShapeToString(shape1), ShapeToString(shape2)};
     }
-    std::vector<uint32_t> output_shape(std::max(shape1.size(), shape2.size()));
+    std::vector<size_t> output_shape(std::max(shape1.size(), shape2.size()));
     for (int i = 0; i < std::max(shape1.size(), shape2.size()); i++) {
         int index1 = static_cast<int>(shape1.size()) - i - 1;
         int index2 = static_cast<int>(shape2.size()) - i - 1;
-        uint32_t dim1 = CheckIndex(index1, shape1.size()) ? shape1[index1] : 1;
-        uint32_t dim2 = CheckIndex(index2, shape2.size()) ? shape2[index2] : 1;
+        size_t dim1 = CheckIndex(index1, shape1.size()) ? shape1[index1] : 1;
+        size_t dim2 = CheckIndex(index2, shape2.size()) ? shape2[index2] : 1;
         output_shape[output_shape.size() - i - 1] = std::max(dim1, dim2);
     }
     return output_shape;
 }
 
 
-std::vector<uint32_t> RemoveDim(const std::vector<uint32_t>& shape, int dim, bool keepdims) {
-    std::vector<uint32_t> new_shape(shape.size() - (1 - keepdims));
+std::vector<size_t> RemoveDim(const std::vector<size_t>& shape, int dim, bool keepdims) {
+    std::vector<size_t> new_shape(shape.size() - (1 - keepdims));
     if (dim < 0) {
         dim += shape.size();
     }
